% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerest.R
\name{powerest}
\alias{powerest}
\title{Function that makes power approximation based on simulated data using \code{simrecurprop}}
\usage{
powerest(
  nsims = 1000,
  n = 1000,
  beta = 0,
  gamma = 0,
  mu0,
  Lam0D,
  crate = 0.01,
  accrualtime = 0,
  admincens = 8000,
  alpha = 0.05
)
}
\arguments{
\item{nsims}{Total number of simulations}

\item{n}{Total sample size. Randomisation 1:1}

\item{beta}{Treatment effect on recurrent events}

\item{gamma}{Treatment effect on terminal events}

\item{mu0}{Reference rate for marginal mean. Data frame with times and mu0(times)}

\item{Lam0D}{Reference rate for cumulative hazard of death. Data frame with times and Lam0D(times)}

\item{crate}{Exponential censoring rate during trial}

\item{accrualtime}{Max accrual time. Same time unit as mu0 and Lam0D}

\item{admincens}{Study closure, time of administrative censoring.  Same time unit as mu0 and Lam0D}

\item{alpha}{Significance level. Default is 0.05}
}
\description{
Function that makes power approximation based on simulated data using \code{simrecurprop}
}
\keyword{simpowerrecurrent}
