% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simrecurprop.R
\name{simrecurprop}
\alias{simrecurprop}
\title{Function that simulates a single data set according to the proportional means
and proportional hazards model}
\usage{
simrecurprop(
  n,
  beta = 0,
  gamma = 0,
  mu0,
  Lam0D,
  crate = 0.01,
  accrualtime = 0,
  admincens = 8000
)
}
\arguments{
\item{n}{TOtal sample size. Randomisation 1:1}

\item{beta}{Treatment effect on recurrent events}

\item{gamma}{Treatment effect on terminal events}

\item{mu0}{Reference rate for marginal mean. Data frame with times and mu0(times)}

\item{Lam0D}{Reference rate for cumulative hazard of death. Data frame with times and Lam0D(times)}

\item{crate}{Exponential censoring rate during trial}

\item{accrualtime}{Max accrual time. Same time unit as mu0 and Lam0D}

\item{admincens}{Study closure, time of administrative censoring.  Same time unit as mu0 and Lam0D}
}
\description{
Function that simulates a single data set according to the proportional means
and proportional hazards model
}
\keyword{simpowerrecurrent}
